move register worker
