+++ packages/core/src/modules/airtable/airtable.core.ts
@@ -9,11 +9,11 @@ import {
   Student,
 } from '@oyster/types';
 
+import { registerWorker } from '@/infrastructure/bull/bull';
 import {
   AirtableBullJob,
   type GetBullJobData,
 } from '@/infrastructure/bull/bull.types';
-import { registerWorker } from '@/infrastructure/bull/use-cases/register-worker';
 import { IS_PRODUCTION } from '@/shared/env';
 import { ColorStackError, ErrorWithContext } from '@/shared/errors';
 import { RateLimiter } from '@/shared/utils/rate-limiter';
