+++ packages/core/src/modules/member/profile.worker.ts
@@ -1,7 +1,7 @@
 import { match } from 'ts-pattern';
 
+import { registerWorker } from '@/infrastructure/bull/bull';
 import { ProfileBullJob } from '@/infrastructure/bull/bull.types';
-import { registerWorker } from '@/infrastructure/bull/use-cases/register-worker';
 import { sendProfileViewsNotification } from './use-cases/send-profile-views-notification';
 
 export const profileWorker = registerWorker(
