+++ packages/core/src/modules/onboarding-session/onboarding-session.worker.ts
@@ -1,7 +1,7 @@
 import { match } from 'ts-pattern';
 
+import { registerWorker } from '@/infrastructure/bull/bull';
 import { OnboardingSessionBullJob } from '@/infrastructure/bull/bull.types';
-import { registerWorker } from '@/infrastructure/bull/use-cases/register-worker';
 import { onOnboardingSessionAttended } from './events/onboarding-session-attended';
 
 export const onboardingSessionWorker = registerWorker(
